% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ms.R
\name{theme_ms_line}
\alias{theme_ms_line}
\title{Theme for line plots}
\usage{
theme_ms_line(base_family = "IBMPlexSans", base_size = 12, grid = TRUE)
}
\arguments{
\item{grid}{Whether grid is printed or not, default=TRUE}
}
\description{
This theme relies on the IBM Plex Sans font which should be installed in advance.
For convenience, run import_ps() to install font.
}
\examples{

}
