% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ms.R
\name{theme_ms_scatter}
\alias{theme_ms_scatter}
\title{Theme for scatter plots}
\usage{
theme_ms_scatter(base_family = "IBMPlexSans", base_size = 9,
  plot_title_family = "IBMPlexSans-Bold", plot_title_size = 18,
  subtitle_size = 13, caption_size = 9, grid = TRUE)
}
\arguments{
\item{grid}{Whether grid is printed or not, default=TRUE}
}
\description{
This theme relies on the IBM Plex Sans font which should be installed in advance.
For convenience, run import_ps() to install font.
}
\examples{

}
