% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ms.R
\name{theme_ms}
\alias{theme_ms}
\title{Custom theme for ggplot2}
\usage{
theme_ms(base_family = "IBMPlexSans", base_size = 12, grid = TRUE, dark FALSE, alttf = FALSE)
}
\arguments{
\item{grid}{Whether grid is printed or not, default=TRUE}
\item{dark}{Switch color scheme to dark: default=FALSE}
\item{alttf}{Choose alternative title font, default=FALSE}
}
\description{
This theme relies on the IBM Plex Sans font which should be installed in advance.
For convenience, run import_ps() to install font. Alternatively, Playfair Display font can be used. Please run import_pf() for installation.
}
\examples{

}
